requirements.txt
alembic==1.11.1
fastapi~=0.100.0
numpy==1.25.0
onnxruntime==1.15.1
Pillow==10.0.0
pydantic~=2.0.2
SQLAlchemy==2.0.18
uvicorn==0.22.0
aiogram==3.0.0b8

roboflow~=1.1.0